Update migrations for onapprovedhook, oncompletehook, and ontransithook models
The commit introduces a new migration file changing several fields in the onapprovedhook, oncompletehook, and ontransithook models. Hooks now have a "callback_function" related to the "river.function", a "hook_type" with the choices "Before" and "After", and a "workflow" related to "river.workflow". Additionally, the status field has been updated in the transition and transitionapproval models.